+++ Reports/NumberOfSalesReport.cs
@@ -35,7 +35,7 @@ namespace Nwazet.Commerce.Reports {
             get { return ChartType.Line; }
         }
 
-        public override double ComputeResultForInterval(IList<OrderPart> ordersForInterval) {
+        public override decimal ComputeResultForInterval(IList<OrderPart> ordersForInterval) {
             return ordersForInterval.Count();
         }
     }
